Removed the cache_clear class method from models and, along with it, removed the class instance variable @caches.
There are a couple of reasons for this. The first reason is that cache_clear relies on state data stored in the class and instances of the class can exist in multiple places. For instance, if you have a number of unicorn instances running, each will have different values in the variable and as such there is no guarantee that calling Model.clear_cache will actually clear the cache for the model. If you really need to clear your cache in a specific way, it should be done on the cache itself.

The second is that in its current state, every time a cache is set or fetches, a new key is added to one of the arrays in @caches. This will often lead to a serious memory leak. Although cache_del will remove the key, caches are far more likely to expire than they are to be manually deleted. To make matters worse, new cache keys are simply pushed into the arrays with no checks for uniqueness, so selecting the same model over and over will still cause the arrays to grow.

Unless this feature can be implemented entirely at the driver end with no state data being stored in the class itself, it is too dangerous of a feature to maintain.